refactor: extract third nav retrieval logic to a util function
This commit extracts the third nav retrieval logic into its own util
function so it can be re-used elsewhere.
